requirements.txt
requests==2.25.1
tqdm==4.61.0
pandas==1.2.4
beautifulsoup4==4.9.3
python_dateutil==2.8.1
